Add versions composer command
This adds a command that either lists all the installed versions, or shows a chosen package's version. Hopefully @seldaek has no plans for this one.
Possible next step: add some ANSI highlights to versions, so that the version is displayed in <bold> as opposed to SHA that should be displayed in gray.
Maybe it's worth also adding an option that would hide the SHAs.