Handle when `composer install —no-dev` was used
The `installed.json` could contain an empty array for packages (like this one) that only use `require-dev` and have nothing in `require`. This will address that by merging both the json and lock files.

As a consequence, the FallbackVersionsTest::testValidVersionsWithoutComposerLock() will need to be skipped when the installed.json is empty.