Rewrote package to use `composer-runtime-api:^2.0` and no longer keep an internal generated map
This massively simplifies the whole design, as it allows us to get rid of most of the complexity
of the package. Sadly, due to the API of `Versions::ROOT_PACKAGE_NAME`, we still need to generate
code during installation, and the composer plugin needs to be maintained long-term, but overall
we can deprecate its usage, and endorse moving to `composer/composer:^2.0.0`.