Add PHP 8.0.* support
This patch explicitly adds `php:~8.0.0` support along with `php:^7.4.7`.

`phpunit/phpunit` had to also be upgraded, so that it does not contain symbols
that clash with the `php:8.0.0` parser BC breaks.

To do so, we had to disable `infection/infection` reporting for now, as the
mutation testing setup is not yet capable of working with `phpunit/phpunit:^9.3`.

In addition to that, we now actively replace `composer/package-versions-deprecated`,
which is constantly causing issues in installation, and which is fully replaced by
this specific package. If you run `composer/composer:^2.0` and `php:8.0.0`, there is
no reason to use `composer/package-versions-deprecated` anyway.

Please note that the support for PHP is limited to `php:^7.4.7 || php:~8.0.0`: `php:8.1.0`
is **NOT** supported by this release, as the upstream ecosystem is just too unstable
and regularly breaks BC in multiple ways. Once `php:8.1.0-rc2` is out (possibly next year),
work will be done to support that too.